+++ wkcuber/wkcuber/check_equality.py
@@ -9,7 +9,6 @@ import numpy as np
 
 from wkcuber.mag import Mag
 from .utils import (
-    add_silent_flag,
     open_wkw,
     WkwDatasetInfo,
     add_distribution_flags,
@@ -53,7 +52,6 @@ def create_parser() -> ArgumentParser:
         default=None,
     )
 
-    add_silent_flag(parser)
     add_distribution_flags(parser)
 
     return parser